Add duck pattern to redux directory
