+++ src/components/redux/book-redux/BookSlice.js
@@ -1,6 +1,7 @@
 import { createSlice } from '@reduxjs/toolkit';
 
 // Define the slice of state for books
+
 const booksSlice = createSlice({
   name: 'books',
   initialState: { books: [] },
@@ -15,19 +16,6 @@ const booksSlice = createSlice({
     }),
   },
 });
-// Define the slice of state for categories
-const categoriesSlice = createSlice({
-  name: 'categories',
-  initialState: { categories: ['Under construction'] },
-  reducers: {
-    checkStatus: () => 'Under construction',
-  },
-});
-
-// Export the reducers for use in the store
+// Export the  book reducers for use in the store
 export const { addBook, removeBook } = booksSlice.actions;
-export const { checkStatus } = categoriesSlice.actions;
-
-// Export the slice of state for use in the store
 export const booksReducer = booksSlice.reducer;
-export const categoriesReducer = categoriesSlice.reducer;
